requirements.txt
Django==3.2.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
djangorestframework==3.12.2
django-bootstrap-v5==1.0.11
django_serverside_datatable==2.1.0
drf_yasg==1.21.7